requirements.txt
asgiref==3.5.2
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.50
platformdirs==2.5.2
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
Django>=3.0,<4.0
psycopg2-binary>=2.8
celery
redis
django-redis